requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Admin==1.6.1
Werkzeug==2.3.7
pjsua2==2.13
faster-whisper==0.9.0
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0
coqui-tts==0.20.0
espeak-ng==0.1.0
soundfile==0.12.1
numpy==1.24.3
librosa==0.10.1
pydub==0.25.1
redis==5.0.1
celery==5.3.4 